finish file themselves
