+++ client/src/views/pages/recommendation/OwnRecommendation.js
@@ -1,8 +1,15 @@
 import React, { useState, useEffect } from 'react'
-import RecomPost from './RecomPost'
+import CareerBlock from '../../components/career/CareerBlock'
+import Masonry from 'react-masonry-css'
 import axios from 'axios'
 const OwnRecommendation = () => {
   const [data, setData] = useState([])
+  const breakpointColumnsObj = {
+    default: 2,
+    1100: 2,
+    700: 1,
+    500: 1,
+  }
   const getData = () => {
     axios
       .get('/api/recommendation/mine')
@@ -18,7 +25,30 @@ const OwnRecommendation = () => {
     getData()
   }, [])
   return (
-    <div className="text-color-black">{data && <RecomPost data={data} setData={setData} />}</div>
+    <div className="text-color-black">
+      <a href="/#/addRecommendation">
+        <div className="d-flex justify-content-center add" width="100%">
+          +
+        </div>
+      </a>
+      {data.length !== 0 && (
+        <Masonry
+          breakpointCols={breakpointColumnsObj}
+          className="my-masonry-grid"
+          columnClassName="my-masonry-grid_column"
+          columnAttrs={{
+            className: 'should be overridden',
+            'data-test': '',
+            style: { '--test': 'test' },
+          }}
+          style={{ display: 'flex' }}
+        >
+          {data.map((post, i) => (
+            <CareerBlock post={post} setData={setData} index={i} key={i} />
+          ))}
+        </Masonry>
+      )}
+    </div>
   )
 }
